+++ src/session.rs
@@ -64,6 +64,8 @@ pub async fn authtoken(authtoken: String) {
 }
 
 /// The builder for an ngrok session.
+///
+/// @group Sessions
 #[napi]
 #[allow(dead_code)]
 #[derive(Default)]
@@ -391,6 +393,8 @@ impl NgrokSessionBuilder {
 }
 
 /// An ngrok session.
+///
+/// @group Sessions
 #[napi(custom_finalize)]
 pub(crate) struct NgrokSession {
     #[allow(dead_code)]
@@ -463,6 +467,9 @@ impl ObjectFinalize for NgrokSession {
     }
 }
 
+/// Container for UpdateRequest information.
+///
+/// @group Sessions
 #[derive(Clone)]
 #[napi]
 pub struct UpdateRequest {